+++ src/mongo/db/index/skipped_record_tracker.cpp
@@ -150,7 +150,7 @@ Status SkippedRecordTracker::retrySkippedRecords(OperationContext* opCtx,
 
     int logLevel = (resolved > 0) ? 0 : 1;
     LOGV2_DEBUG(23883,
-                logLevel.toInt(),
+                logLevel,
                 "index build: reapplied {resolved} skipped records for index: "
                 "{indexCatalogEntry_descriptor_indexName}",
                 "resolved"_attr = resolved,
